services/discord-bug-triage/requirements.txt
discord.py>=2.3,<3
PyGithub>=2.1,<3
